bitrange::from<>() to works in both directions
Before, bitrange::from<> required that the first argument be less than
or equal to the second argument. Otherwise a compiler error is emitted.

But in many cases, a datasheet or user manual will have the bit fields
start from lower to higher or higher to lower. Sometimes the user
has a preference for one or the other.

Rather than emit a compiler error, we can simply, do the right thing and
create a bitrange that represents the correct bit position and width
based on which value is larger.

For example these two statements are equivalant

```C++
constexpr auto field1 = xstd::bitrange::from<4, 6>();
constexpr auto field2 = xstd::bitrange::from<6, 4>();

static_assert(field1 == field2);
```

Also added a == comparison operator for bitrange