Fix xstd::bitrange::from() range calculation
::from() should be able to take a bit range from <A, B> and <B, A> where
A is greater than B and return the same exact bit range. This is there
to make it easier to follow user manuals and their method of writting
out bit ranges (high to low or low to high).

Previous example resulted in a negative number for the bitwidth which
broke the insert/extract functions.